# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth Info Service API Documentation",
        "title": "Auth Info API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/document/generate-pdf": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据模板名称和数据生成 PDF 文件。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "生成 PDF 文档",
                "parameters": [
                    {
                        "description": "生成 PDF 参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apipb.GeneratePDFRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF 文件",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "404": {
                        "description": "模板不存在",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            }
        },
        "/document/generate-word": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据模板名称和数据生成 Word 文档文件。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "生成 Word 文档",
                "parameters": [
                    {
                        "description": "生成 Word 参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apipb.GenerateWordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Word 文档文件",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "404": {
                        "description": "模板不存在",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apipb.GeneratePDFRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/structpb.Struct"
                },
                "template_name": {
                    "type": "string"
                }
            }
        },
        "apipb.GenerateWordRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/structpb.Struct"
                },
                "template_name": {
                    "type": "string"
                }
            }
        },
...
        "apipb.OperationReply": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "structpb.Struct": {
            "type": "object",
            "properties": {
                "fields": {
                    "description": "Unordered map of dynamically typed values.",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/structpb.Value"
                    }
                }
            }
        },
        "structpb.Value": {
            "type": "object",
            "properties": {
                "kind": {
                    "description": "The kind of value.\n\nTypes that are valid to be assigned to Kind:\n\n\t*Value_NullValue\n\t*Value_NumberValue\n\t*Value_StringValue\n\t*Value_BoolValue\n\t*Value_StructValue\n\t*Value_ListValue"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}